Revert "Location client API: ues only Gnss stack for time-based session"
This reverts commit 8bed8f7a04673026ce0df2eb3b19b284fb4ce584.

CRs-Fixed: 2912212

+++ location_hal_daemon/LocHalDaemonClientHandler.cpp
@@ -60,11 +60,11 @@ void LocHalDaemonClientHandler::updateSubscription(uint32_t mask) {
         onCollectiveResponseCallback(count, errs, ids);
     };
 
-    if (mSubscriptionMask & E_LOC_CB_DISTANCE_BASED_TRACKING_BIT) {
-        mCallbacks.trackingCb = [this](Location location) {
-            onTrackingCb(location);
-        };
-    }
+    // update optional callback - following four callbacks can be controlable
+    // tracking
+    mCallbacks.trackingCb = [this](Location location) {
+        onTrackingCb(location);
+    };
 
     // batching
     if (mSubscriptionMask & E_LOC_CB_BATCHING_BIT) {
@@ -94,7 +94,7 @@ void LocHalDaemonClientHandler::updateSubscription(uint32_t mask) {
     }
 
     // location info
-    if (mSubscriptionMask & (E_LOC_CB_GNSS_LOCATION_INFO_BIT | E_LOC_CB_SIMPLE_LOCATION_INFO_BIT)) {
+    if (mSubscriptionMask & E_LOC_CB_GNSS_LOCATION_INFO_BIT) {
         mCallbacks.gnssLocationInfoCb = [this](GnssLocationInfoNotification notification) {
             onGnssLocationInfoCb(notification);
         };
@@ -689,7 +689,7 @@ void LocHalDaemonClientHandler::onTrackingCb(Location location) {
     LOC_LOGd("--< onTrackingCb");
 
     if ((nullptr != mIpcSender) &&
-            (mSubscriptionMask & E_LOC_CB_DISTANCE_BASED_TRACKING_BIT)) {
+        (mSubscriptionMask & (E_LOC_CB_TRACKING_BIT | E_LOC_CB_SIMPLE_LOCATION_INFO_BIT))) {
         // broadcast
         string pbStr;
         LocAPILocationIndMsg msg(SERVICE_NAME, location, &mService->mPbufMsgConv);
@@ -808,34 +808,19 @@ void LocHalDaemonClientHandler::onGnssLocationInfoCb(GnssLocationInfoNotificatio
     std::lock_guard<std::mutex> lock(LocationApiService::mMutex);
     LOC_LOGd("--< onGnssLocationInfoCb");
 
-    if ((nullptr != mIpcSender) && (mSubscriptionMask &
-            (E_LOC_CB_GNSS_LOCATION_INFO_BIT | E_LOC_CB_SIMPLE_LOCATION_INFO_BIT))) {
-        bool rc = false;
+    if ((nullptr != mIpcSender) &&
+        (mSubscriptionMask & E_LOC_CB_GNSS_LOCATION_INFO_BIT)) {
         string pbStr;
-        if (mSubscriptionMask & E_LOC_CB_GNSS_LOCATION_INFO_BIT) {
-            LocAPILocationInfoIndMsg msg(SERVICE_NAME, notification, &mService->mPbufMsgConv);
-            if (msg.serializeToProtobuf(pbStr)) {
-                rc = sendMessage(pbStr.c_str(), pbStr.size(), msg.msgId);
-                // purge this client if failed
-                if (!rc) {
-                    LOC_LOGe("failed rc=%d purging client=%s", rc, mName.c_str());
-                    mService->deleteClientbyName(mName);
-                }
-            } else {
-                LOC_LOGe("LocAPILocationInfoIndMsg serializeToProtobuf failed");
+        LocAPILocationInfoIndMsg msg(SERVICE_NAME, notification, &mService->mPbufMsgConv);
+        if (msg.serializeToProtobuf(pbStr)) {
+            bool rc = sendMessage(pbStr.c_str(), pbStr.size(), msg.msgId);
+            // purge this client if failed
+            if (!rc) {
+                LOC_LOGe("failed rc=%d purging client=%s", rc, mName.c_str());
+                mService->deleteClientbyName(mName);
             }
         } else {
-            LocAPILocationIndMsg msg(SERVICE_NAME, notification.location, &mService->mPbufMsgConv);
-            if (msg.serializeToProtobuf(pbStr)) {
-                rc = sendMessage(pbStr.c_str(), pbStr.size(), msg.msgId);
-                // purge this client if failed
-                if (!rc) {
-                    LOC_LOGe("failed rc=%d purging client=%s", rc, mName.c_str());
-                    mService->deleteClientbyName(mName);
-                }
-            } else {
-                LOC_LOGe("LocAPILocationIndMsg serializeToProtobuf failed");
-            }
+            LOC_LOGe("LocAPILocationInfoIndMsg serializeToProtobuf failed");
         }
     }
 }
